requirements.txt
# Model export
torch>=2.0
openunmix>=1.3
onnx>=1.14

# Inference + benchmark
onnxruntime>=1.15
numpy>=1.24
scipy>=1.11
soundfile>=0.12
mir-eval>=0.7
pyyaml>=6.0
